Move states to pages
