+++ src/tests/ui/pages/state-needs-auth-info.test.ts
@@ -1,7 +1,7 @@
 import "@testing-library/jest-dom";
 import { fireEvent, render, screen } from "@testing-library/svelte";
 import { describe, test, expect, vi } from "vitest";
-import StateNeedsAuthInfo from "../../../ui/states/state-needs-auth-info.svelte";
+import StateNeedsAuthInfo from "../../../ui/pages/state-needs-auth-info.svelte";
 import { SDKEventName } from "../../../behavioural-events/sdk-events";
 import { createEventsTrackerMock } from "../../mocks/events-tracker-mock-provider";
 import { eventsTrackerContextKey } from "../../../ui/constants";
